1. Cambiado el tipo de las instrucciones de unsigned int a unsigned long 2. Cambiado '.text' por '.code' 3. Cambiado el carácter de comentario de '#' a ';'

+++ label_list.c
@@ -14,7 +14,7 @@ void free_label_list(struct LABEL_LIST *list) {
     }
 }
 
-int add_to_list(struct LABEL_LIST *list, char *label, unsigned int num_ins, int allow_duplicates) {
+int add_to_list(struct LABEL_LIST *list, char *label, unsigned long num_ins, int allow_duplicates) {
     struct LABEL_LIST_NODE *node, *aux;
     if (!(list->init)) {
         list->init = (struct LABEL_LIST_NODE *)malloc(sizeof *(list->init));
